+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# Version 1.34 (not yet released)
+# Version 1.34
 
 * Fixed issue with filtering on state/province custom fields.
 * Fixed issue with filtering on multi value fields (fields which holds multiple values separated by the VALUE SEPARATOR symbol).